requirements.txt
altgraph==0.17.4
anyio==4.9.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
dnspython==2.7.0
evdev==1.9.2
eventlet==0.39.1
Flask==3.1.0
Flask-SocketIO==5.5.1
Flask-Sockets==0.2.1
gevent==25.4.2
gevent-websocket==0.10.1
greenlet==3.2.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
metar-taf-parser-mivek==1.9.0
packaging==25.0
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
PyMouse==1.0
pynput==1.8.1
python-engineio==4.12.0
python-socketio==5.13.0
python-xlib==0.33
requests==2.32.3
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
wsproto==1.2.0
zope.event==5.0
zope.interface==7.2
